Move platform specific functions to own files
